+++ test/scenes.js
@@ -62,7 +62,7 @@ test('Restarting the app preserves the default sources', async t => {
   const client = await getClient();
   const app = t.context.app;
   const sceneName = 'Coolest Scene Ever';
-  const appService = client.getResource('AppService');
+  const sceneCollectionsService = client.getResource('SceneCollectionsService');
 
   await addScene(t, sceneName);
 
@@ -70,7 +70,7 @@ test('Restarting the app preserves the default sources', async t => {
   t.true(await app.client.isExisting(`div=${sceneName}`));
 
   // reload config
-  await appService.loadConfig('scenes');
+  await sceneCollectionsService.load(sceneCollectionsService.collections[0].id);
 
   await focusMain(t);
   await selectScene(t, sceneName);
